feat: replace list with table and more details for units that belong to this issuance
