+++ src/components/blocks/ProjectIssuanceDetails.js
@@ -1,8 +1,5 @@
-import React, { useMemo } from 'react';
+import React from 'react';
 import { FormattedMessage } from 'react-intl';
-import { useNavigate } from 'react-router-dom';
-import { useSelector } from 'react-redux';
-import styledComponents from 'styled-components';
 
 import { Body } from '..';
 import {
@@ -11,30 +8,9 @@ import {
   StyledItem,
 } from '.';
 import { detailsViewData } from '../../utils/functionUtils';
-import { getMyOrgUid } from '../../utils/getMyOrgUid';
-import { MagnifyGlassIcon } from '../icons';
-
-const StyledCursor = styledComponents('div')`
-  cursor: pointer;
-`;
+import { DetailedViewIssuanceUnitTable } from './DetailedViewIssuanceUnitTable';
 
 const ProjectIssuanceDetails = ({ data, stagingData, changeColor }) => {
-  const { units } = useSelector(store => store.climateWarehouse);
-  const navigate = useNavigate();
-  const { organizations } = useSelector(store => store.climateWarehouse);
-  const myOrgUid = getMyOrgUid(organizations);
-
-  const unitsBelongingToThisIssuance = useMemo(
-    () =>
-      units?.reduce((accumulator, currentUnit) => {
-        if (currentUnit.issuanceId === data.id) {
-          return [...accumulator, currentUnit.warehouseUnitId];
-        }
-        return accumulator;
-      }, []),
-    [units, data],
-  );
-
   return (
     <StyledDetailedViewTabItem>
       <div style={{ width: '60%' }}>
@@ -98,30 +74,7 @@ const ProjectIssuanceDetails = ({ data, stagingData, changeColor }) => {
                 changeColor,
               )}
           </StyledItem>
-          {data && (
-            <StyledItem>
-              <Body size="Bold" width="100%">
-                <FormattedMessage id="units-belonging-to-issuance" />
-              </Body>
-              {unitsBelongingToThisIssuance?.length > 0 &&
-                unitsBelongingToThisIssuance.map(unitId => (
-                  <StyledCursor key={unitId}>
-                    <Body
-                      onClick={() =>
-                        navigate(
-                          `/units?orgUid=${myOrgUid}&myRegistry=true&unitId=${unitId}`,
-                        )
-                      }
-                      color="#1890ff"
-                    >
-                      {unitId}
-                      <MagnifyGlassIcon height="15" width="30" />
-                    </Body>
-                  </StyledCursor>
-                ))}
-              {unitsBelongingToThisIssuance?.length === 0 && '---'}
-            </StyledItem>
-          )}
+          {data && <DetailedViewIssuanceUnitTable issuance={data} />}
         </StyledDetailedViewTab>
       </div>
     </StyledDetailedViewTabItem>
